Refactors MainViewController into separate view controllers. (#145)
* Refactors everything into separate view controllers.

Using proper view controller containment. Everything thing should work the way it did before, more or less. I made a few tweaks here and there.

* Uses NSTabViewController for the sample view controller.

Why reinvent the wheel?

* Fixes search not returning the correct results initially.

* Fixes split view sizing issues.

* Uses proper method of making SampleListViewControllerDelegate class-bound

* Renames SampleCollectionHeaderView to SampleCollectionSectionHeaderView

The change clarifies that the view is a section header view, not a view displayed on top of the collection view.

* Makes it even more explicit that SampleCollectionSectionHeaderView is a section header view.